p_value complete

+++ facs/prob.h
@@ -3,4 +3,4 @@ extern double get_sigma (long long num_hit, double prob);
 double erf (double x);
 double pdf (double x, double mu, double sigma);
 extern double cdf (double x, double mu, double sigma);
-extern int prob_suggestion (int k_mer);
+extern double prob_suggestion (int k_mer);
